api/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.9
